+++ src/main/java/net/infstudio/goki/common/stat/tool/StatSwordsmanship.java
@@ -21,7 +21,7 @@ public class StatSwordsmanship extends ToolSpecificStat {
     }
 
     @Override
-    public float getBonus(int level) {
+    public double getBonus(int level) {
         return getFinalBonus((float) Math.pow(level, 1.0895D) * 0.03F);
     }
